requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pandas
rdkit-pypi
streamlit